Add files via upload
This is perceptron Learning algorithm.
Perceptron learning is a type of supervised learning algorithm used for binary classification problems. It is based on a single-layer artificial neural network known as a perceptron.

The perceptron takes multiple inputs and applies weights to each input. The weighted inputs are then passed through an activation function, which produces an output. The output is compared to the desired output, and the weights are adjusted accordingly to reduce the error between the output and the desired output.

The perceptron learning algorithm works by updating the weights in the direction of the misclassified samples. This process is repeated until all samples are correctly classified or a stopping criterion is met.

Perceptron learning is a simple and effective algorithm for linearly separable problems, but it can struggle with problems that are not linearly separable. However, variations of the perceptron learning algorithm, such as the multilayer perceptron, have been developed to handle more complex problems.

This code is done from scratch and also checks whether the points are linearly separable or not.